Updated libutils version

+++ libutils/numutils.hpp
@@ -6,6 +6,7 @@ Last update: 2026 Mar 26 */
 #define NUMUTILS_HPP
 
 #include <algorithm>
+#include <cstdlib>
 #include <cstdint>
 #include <iomanip>
 #include <sstream>
@@ -21,7 +22,7 @@ inline std::string groupDigits(long long value, char separator = ',') {
   bool negative = value < 0;
   std::string s = std::to_string(std::llabs(value));
 
-  for (int i = s.length() - 3; i > 0; i -= 3)
+  for (size_t i = s.length() - 3; i > 0; i -= 3)
     s.insert(i, 1, separator);
 
   if (negative)
@@ -46,7 +47,7 @@ template <typename T> inline std::string fixed(T value, int precision) {
  */
 inline std::string zeroPad(long long value, size_t width) {
   std::ostringstream ss;
-  ss << std::setw(width) << std::setfill('0') << value;
+  ss << std::setw((int)width) << std::setfill('0') << value;
   return ss.str();
 }
 
@@ -255,7 +256,7 @@ inline std::string time_ms(uint64_t ms) {
     return ss.str();
   }
 
-  double seconds = ms / 1000.0;
+  double seconds = static_cast<double>(ms) / 1000.0;
 
   if (seconds < 60) {
     ss << std::fixed << std::setprecision(2) << seconds << "s";
